ref(mypy): Rename sort_stronger_modules to sort_weaklist
The new name matches the pre-commit hook id (sort-mypy-weaklist) and the
rest of the weaklist tooling (check_stronglist, prevent_weaklist_additions).

A re-export shim is kept at the old module path so getsentry's
.pre-commit-config.yaml keeps working until it is updated to call the new
name. Once getsentry is updated, a follow-up will remove the shim.
